requirements.txt
playwright==1.41.0
pytest==8.0.0
openai==1.12.0
python-dotenv==1.0.1
pytest-playwright==0.4.4